Internal change
PiperOrigin-RevId: 377131555
